alphatrader/requirements.txt
# AlphaTrader Dependencies
# Phase 0: Neural Network Imitation Learning System

# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data fetching
python-binance>=1.0.17
requests>=2.28.0
aiohttp>=3.8.0

# Technical Analysis
ta-lib>=0.4.25  # Requires TA-Lib C library installed
pandas-ta>=0.3.14b  # Pure Python alternative

# Machine Learning
torch>=2.0.0
scikit-learn>=1.2.0
hmmlearn>=0.3.0  # Hidden Markov Models

# Reinforcement Learning
gymnasium>=0.28.0  # OpenAI Gym successor

# Monitoring & Logging
tensorboard>=2.12.0
tqdm>=4.65.0
rich>=13.0.0  # Beautiful terminal output

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
# SQLite is built-in, no extra deps needed

# Discord notifications
discord-webhook>=1.1.0

# GPU monitoring
pynvml>=11.5.0  # NVIDIA GPU monitoring
GPUtil>=1.4.0

# Development
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.22.0
